portfolio/requirements.txt
ansible==4.3.0
ansible-core==2.11.3
astroid==2.2.5
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2019.6.16
cffi==1.14.6
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
croniter==0.3.37
cryptography==3.4.7
dask==2021.4.0
distributed==2021.4.0
django-widget-tweaks==1.4.8
docker==4.4.4
Flask==1.1.2
Flask-PyMongo==2.3.0
fsspec==0.8.7
google==3.0.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
HeapDict==1.0.1
httplib2==0.19.1
idna==2.8
importlib-metadata==3.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.1
locket==0.2.1
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-oneofschema==2.1.0
mccabe==0.6.1
mongoengine==0.23.0
msgpack==1.0.2
mypy-extensions==0.4.3
mysql-connector==2.2.9
natsort==7.1.1
oauthlib==3.1.0
packaging==20.9
partd==1.1.0
pendulum==2.1.2
prefect==0.14.14
protobuf==3.15.7
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pylint==2.3.1
pymongo==3.10.1
pyparsing==2.4.7
python-box==5.3.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
resolvelib==0.5.4
rsa==4.7.2
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2.1
tabulate==0.8.9
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==16.7.9
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.11.2
zict==2.0.0
zipp==3.4.0
